Add viewModels and refactor MainFragment
