+++ app/src/main/java/com/rafaelbandeeira/chucknorrisjokes/network/Jokes.kt
@@ -1,4 +1,4 @@
-package com.rafaelbandeeira.chucknorrisjokes
+package com.rafaelbandeeira.chucknorrisjokes.network
 
 data class Jokes (
     var icon_url: String?,
