+++ app/src/main/res/layout/activity_main.xml
@@ -17,6 +17,6 @@
         app:layout_constraintBottom_toBottomOf="parent"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        tools:context=".MainFragment" />
+        tools:context=".fragments.MainFragment" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
